Avoiding unloading the Freetype library and typeface during texture_font_load_glyphs() only to immediately load it again to calculate the kerning.
Explanation: texture_font_generate_kerning() is called only in one place in the entire library, and that's inside texture_font_load_glyphs().  Moving it to before the cleanup of library and face allows us to avoid a costly and completely pointless unload followed by reload.  texture_font_generate_kerning() is not declared in the headers so there is no risk of it being called elsewhere, so this change has no risk of breaking anything in any existing code.  A pointer to the existing face is passed along instead.