Add screenshot functionality to all demos
Adding an optional `--screenshot <file-path>` parameter to all demos.
When passing this parameter the first frame of the corresponding demo is
rendered and saved as Targa file to the given file path.  After this the
demo exits.

This commit enables the demos to be tested in a more automated fashion.