Universal support for RGBA texture atlas
This adds support for using one RGBA (4-channel) texture atlas, even for greyscale (1-channel) fonts, so that one can use the same texture atlas for ordinary text and emojis, simplifying implementation and reducing the number of separate draw calls.

When a greyscale font is encountered, it is copied into the texture atlas with white pixels, which can be later colored in the shader if desired.