makefont: Fix invalid C data structures declarations
The data structures generated by makefont are not valid, and may fail
to compile depending on compiler flags:

  error: missing braces around initializer [-Werror=missing-braces]
  | texture_glyph_t font_glyph_00000020 =   {32, 2, 2, 0, 0, 9.640625f
  |                                         ^

The texture_glyph_t::kerning field is a double array, so it is invalid
to initialize it as an integer, "0". When a glyph has no kerning
pairs, initialize the "kerning" field to "{{0}}".

Remove the "texture_glyph_0x100_t" structure and incorporate the
glyphs pointer array directly into texture_font_t.